Fix: Repository 매개변수 type 수정

+++ src/main/java/com/example/mind_laundromat/cbt/repository/DiaryRepository.java
@@ -3,5 +3,5 @@ package com.example.mind_laundromat.cbt.repository;
 import com.example.mind_laundromat.cbt.entity.Diary;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DiaryRepository extends JpaRepository<Diary, Integer> {
+public interface DiaryRepository extends JpaRepository<Diary, Long> {
 }
